Finished the first draft of the new scene strategy.
This has only had the most basic tests, and so still needs to be put
through it's paces. It does resemble the original a bit, but this one
still has some kinks.

Uses C++11, so I'll need to keep that in mind. I'll need to determine
which version fo the scene strategy works best.